tests/wpt/tests/webauthn/conditional-mediation.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<title>Conditional Mediation tests</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src=helpers.js></script>
<body></body>
<script>
"use strict";

// Test that conditional mediation (get) is supported.
virtualAuthenticatorPromiseTest(async t => {
  assert_own_property(window.PublicKeyCredential, "isConditionalMediationAvailable");
  assert_true(await window.PublicKeyCredential.isConditionalMediationAvailable());
}, {
  protocol: "ctap2",
  hasResidentKey: true,
  hasUserVerification: true,
  transport: "internal",
}, "Conditional mediation supported"
);


// Test that conditional create is supported.
virtualAuthenticatorPromiseTest(async t => {
  assert_own_property(window.PublicKeyCredential, "getClientCapabilities");
  let capabilities = await window.PublicKeyCredential.getClientCapabilities();
  assert_true(typeof (capabilities["conditionalCreate"]) == "boolean");
  assert_true(capabilities["conditionalCreate"]);
}, {
  protocol: "ctap2",
  hasResidentKey: true,
  hasUserVerification: true,
  transport: "internal",
}, "Conditional create supported");

</script>
